【bug修复】: 修复删除云盘文件时minio中文件无法删除bug

+++ api/blog-file/src/main/java/com/blog/file/service/impl/UploadFileServiceImpl.java
@@ -83,7 +83,7 @@ public class UploadFileServiceImpl implements UploadFileService {
             fileCategoryData.setFileName(fileName);
             fileCategoryData.setFileCategoryId(categoryId);
             fileCategoryData.setFileUrl(fileUrl);
-            fileCategoryData.setFileSize((int) uploadVo.getFile().getSize());
+            fileCategoryData.setFileSize(uploadVo.getFile().getSize());
             fileCategoryData.setFileStatus(0);
             fileCategoryData.setFileType(fileType);
             if (FileTypeEnum.getTypeEnumByFileType(fileType).getFileType() == 3) {
